Move ClientIdentifier to own mod
