+++ src/lib.rs
@@ -7,4 +7,5 @@
 mod codecs;
 mod packet;
 mod client;
+mod client_identifier;
 mod error;
